examgpt-backend/examgpt_backend/requirements.txt
requests ~= 2.32.3
tiktoken ~= 0.7.0
python-dotenv ~= 1.0.1
loguru ~= 0.7.2
pydantic-settings ~= 2.3.4
python-telegram-bot ~= 21.4
langchain ~= 0.2.6
langchain-core ~= 0.2.11
langchain-community ~= 0.2.6
openai ~= 1.35.10
langchain-openai ~= 0.1.14
codenamize ~= 1.2.3
pymupdf ~= 1.24.9
aws-lambda-powertools ~= 2.43.0
email-validator ~= 2.2.0
# unstructured[pdf] ~= 0.14.10